+++ src/Http/ResponseBuilderInterface.php
@@ -6,6 +6,9 @@ use Psr\Http\Message\ResponseInterface;
 
 /**
  * Builds a PSR-7 Response
+ *
+ * @codeCoverageIgnore
+ * @deprecated
  */
 interface ResponseBuilderInterface
 {